small chng

+++ src/print_matr.c
@@ -1,5 +1,7 @@
 
-void	print_matr(float **matr, int8_t matr_x, int8_t matr_y)
+#include "../includes/filler_header.h"
+
+void	print_matr(int32_t **matr, int8_t matr_x, int8_t matr_y)
 {
 	int i = 0;
 	int	j;
